TST: Move and update TestRobinBoundaryCondition
